try rewriting logical plan
